// Client/companeros-de-viaje-app/src/app/core/interfaces/groupResp.interface.ts
// Generated by https://quicktype.io

export interface GroupResponse {
  id: number;
  destination: string;
  departureDate: string;
  returnDate: string;
  itinerary: string;
  budget: number;
  owner: Owner;
  interests: string[];
  minimumNumberOfMembers: number;
  travelers: Owner[];
  totalTravelers: number;
}

export interface Owner {
  username: string;
  name: string;
  profilePicture: null;
}